+++ pkg/esv/caller.go
@@ -5,6 +5,7 @@
 package esv
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -49,7 +50,7 @@ func (c Client) UserAgentVersion() string {
 	return fmt.Sprintf("go-esv-api/%s", Version)
 }
 
-func (c Client) MakeRequest(path string, os []Option) (http.Request, error) {
+func (c Client) MakeRequest(path string, os []Option) (*http.Request, error) {
 	req := http.Request{
 		Method: "GET",
 	}
@@ -62,7 +63,7 @@ func (c Client) MakeRequest(path string, os []Option) (http.Request, error) {
 
 	pathURL, err := url.Parse(path)
 	if err != nil {
-		return req, fmt.Errorf("error parsing constructed path and query: %w", err)
+		return &req, fmt.Errorf("error parsing constructed path and query: %w", err)
 	}
 
 	requestURL := c.BaseURL.ResolveReference(pathURL)
@@ -72,18 +73,24 @@ func (c Client) MakeRequest(path string, os []Option) (http.Request, error) {
 	req.Header.Add("Authorization", "Token "+c.Token)
 	req.Header.Add("User-Agent", c.UserAgent)
 
-	return req, nil
+	return &req, nil
 }
 
 // CallEndpoint is a generic method for making API calls at the endpoint. This
 // method is exposed to provide flexibility, but should not normally be used.
-func (c Client) CallEndpoint(path string, o []Option, r interface{}) error {
+func (c Client) CallEndpoint(
+	ctx context.Context,
+	path string,
+	o []Option,
+	r interface{},
+) error {
 	req, err := c.MakeRequest(path, o)
 	if err != nil {
 		return fmt.Errorf("error building request: %w", err)
 	}
 
-	res, err := c.Client.Do(&req)
+	req = req.WithContext(ctx)
+	res, err := c.Client.Do(req)
 	if err != nil {
 		return fmt.Errorf("error making request: %w", err)
 	}